Add files via upload
Should be the most updated versions of everything which include: g2phase4zip which has the code for making the accounts done by Alice and Ali. Then we also have the phase 4 docx and UserAccountTests.zip which were made by Aliza and Amy. Finally me and Amanda worked on the Figma link and the screenshots.